Change background color of body to #1F2937
When background color was white, scrolling up on toolbar will cause a
the white background to show up. Background color was changed to #1F2937
to match background color of toolbar so scrolling up wil not reveal a
white background.

Default background color was changed to #1F2937 so background of info
section was changed to white.